+++ test/matchers/java/JavaBlockMicrogrammarTest.ts
@@ -20,7 +20,7 @@ describe("JavaBlock microgrammars", () => {
  * Note that these grammars aren't meant to be realistic
  */
 
-export const JAVA_IDENTIFIER = /^[A-Za-z][a-zA-Z0-9]+/;
+export const JAVA_IDENTIFIER = /[A-Za-z][a-zA-Z0-9]+/;
 
 const METHOD_GRAMMAR = Microgrammar.fromDefinitions<ChangeControledMethod>({
     _visibilityModifier: "public",